# swagger.yaml
openapi: 3.0.0
info:
  description: This is Woleet.ID Server API documentation.
  version: 1.0.4
  title: Woleet.ID Server
  contact:
    email: contact@woleet.com
  license:
    name: MIT
security:
  - CookieAuth: []
paths:
...
  /discover/users:
    get:
      operationId: discoverUsers
      summary: Get all users matching a search string.
      tags:
        - Discovery
      description: >
        Use this endpoint to get all users whose `email`, `username`, `x500CommonName`, `x500Organization` or `x500OrganizationalUnit` contains the search string.
      parameters:
        - $ref: '#/components/parameters/UserSearchQueryParameter'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UserDiscoArray'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
  /discover/user/{pubKey}:
    get:
      operationId: discoverUserByPubKey
      summary: Get the user associated to a public key.
      tags:
        - Discovery
      description: >
        Use this endpoint to get the user owning a given public key.
      parameters:
        - $ref: '#/components/parameters/PubKeyParameter'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UserDisco'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
  /discover/keys/{userId}:
    get:
      operationId: discoverUserKeys
      summary: Get all public keys of a user.
      tags:
        - Discovery
      description: >
        Use this endpoint to get all public keys owned by a given user.
      parameters:
        - $ref: '#/components/parameters/UserIdParameter'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/KeyDiscoArray'
        '400':
          description: Invalid user identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
...
  /user/{userId}:
    get:
      tags:
        - User
      summary: Get a user by his identifier.
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/UserIdParameter'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UserGet'
        '400':
          description: Invalid user identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
    put:
      tags:
        - User
      summary: Update a user.
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserIdParameter'
      requestBody:
        description: User object to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPut'
        required: true
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UserGet'
        '400':
          description: Invalid object supplied / Invalid user identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
    delete:
      tags:
        - User
      summary: Delete a user.
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserIdParameter'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/UserGet'
        '400':
          description: Invalid user identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
...
  /api-token:
    post:
      tags:
        - API Token
      summary: Create an API token.
      description: Create an API token suitable to call the /sign endpoint (this can only be done by an admin).
      operationId: createAPIToken
      requestBody:
        description: API token object to create.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APITokenPost'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APITokenGet'
        '400':
          description: Invalid object supplied.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
  /api-token/{APITokenId}:
    get:
      tags:
        - API Token
      summary: Get an API token by its identifier.
      operationId: getAPITokenById
      parameters:
        - $ref: '#/components/parameters/APITokenIdParameter'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APITokenGet'
        '400':
          description: Invalid API token identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: API token not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
    put:
      tags:
        - API Token
      summary: Update an API token.
      operationId: updateAPIToken
      parameters:
        - $ref: '#/components/parameters/APITokenIdParameter'
      requestBody:
        description: API token object to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/APITokenPut'
        required: true
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APITokenGet'
        '400':
          description: Invalid object supplied / Invalid API token identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: API token not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
    delete:
      tags:
        - API Token
      summary: Delete an API token.
      operationId: deleteAPIToken
      parameters:
        - $ref: '#/components/parameters/APITokenIdParameter'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APITokenGet'
        '400':
          description: Invalid API token identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: API token not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
  /api-token/list:
    get:
      tags:
        - API Token
      summary: List all API tokens.
      operationId: getAllAPITokens
      responses:
        '200':
          description: Succesful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APITokenArray'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
...
  /server-config:
    get:
      tags:
        - Server Config
      summary: Get the server configuration.
      operationId: getServerConfig
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerConfig'
        '400':
          description: Invalid API token identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: API token not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
    put:
      tags:
        - Server Config
      summary: Update the server configuration.
      operationId: updateServerConfig
      requestBody:
        $ref: '#/components/requestBodies/ServerConfig'
      responses:
        '200':
          description: Successful operation.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/ServerConfig'
        '400':
          description: Invalid object supplied / Invalid API token identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '401':
          description: Missing or bad session cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
        '404':
          description: API token not found.
          content:
            application/json:
              schema:
                $ref: '#/components/responses/APIError'
components:
  schemas:
    Identity:
      type: object
      description: User identity provided as a set of X.500 attributes (see https://www.ietf.org/rfc/rfc4519.txt).
      properties:
        commonName:
          type: string
          description: Common name (CN) (2.5.4.3)
          example: "Jim Smith"
        organization:
          type: string
          description: Organization name (O) (2.5.4.10)
          example: "Woleet"
        organizationalUnit:
          type: string
          description: Organizational unit name (OU) (2.5.4.11)
          example: "Sales"
        locality:
          type: string
          description: Locality name (L) (2.5.4.7)
          example: "Rennes"
        country:
          type: string
          description: Country code (C) (2.5.4.6)
          example: "FR"
      required:
        - commonName
...
    FullIdentity:
      allOf:
        - $ref: '#/components/schemas/Identity'
        - type: object
          properties:
            userId:
              type: string
              description: Custom user identifier (UID) (0.9.2342.19200300.100.1.1) must be unique for each user.
              example: "wol.jim-smith.01"
    Password:
      type: string
      format: password
      description: User password.
      example: "nSBa+rV3%2/LpD"
...
    PubKey:
      type: string
      description: Public key (bitcoin address when using BIP39 keys).
      example: "1GChJMuyxvq28F3Uksqf5v7QkxQ4WLQdBh"
    CreatedAt:
      type: integer
      format: int64
      readOnly: true
      description: Date of creation (Unix ms timestamp).
      example: 1529052551419
    UpdatedAt:
      type: integer
      format: int64
      readOnly: true
      description: Date of last modification (Unix ms timestamp).
      example: 1529052551419
    DeletedAt:
      type: integer
      format: int64
      readOnly: true
      description: Date of deletion (Unix ms timestamp) (`null` if not deleted).
      example: 1529052551419
    LastUsed:
      type: integer
      format: int64
      readOnly: true
      description: Date of last usage (Unix ms timestamp).
      example: 1529059167339
    LastLogin:
      type: integer
      format: int64
      readOnly: true
      description: Date of last login (Unix ms timestamp).
      example: 1529050155459
...
    UserRoleEnum:
      type: string
      enum:
       - user
       - admin
      default: "user"
      description: User role.
      example: "user"
    UserStatusEnum:
      type: string
      enum:
      - active
      - blocked
      default: "active"
      description: User status (a `blocked` user cannot sign).
      example: "active"
    KeyStatusEnum:
      type: string
      enum:
      - active
      - blocked
      default: "active"
      description: Key status (a `blocked` key cannot be used to sign).
      example: "active"
    UserBase:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
          description: User identifier (allocated by the server).
          example: "7c42e7e9-aec5-4d56-9a3b-bd55e129aae3"
        email:
          type: string
          format: email
          description: User email (can be used for login).
          example: "john.doe@acme.com"
        username:
          type: string
          description: User name (can be used for login).
          example: "johndoe"
        status:
          $ref: '#/components/schemas/UserStatusEnum'
        role:
          $ref: '#/components/schemas/UserRoleEnum'
        identity:
          $ref: '#/components/schemas/FullIdentity'
        defaultKeyId:
          type: string
          format: uuid
          description: Identifier of the default key to use for this user.
          example: "c7c6e0de-2acb-4311-80b4-17dbf0b76806"
    UserGet:
      allOf:
        - $ref: '#/components/schemas/UserBase'
        - type: object
          properties:
            createdAt:
              $ref: '#/components/schemas/CreatedAt'
            updatedAt:
              $ref: '#/components/schemas/UpdatedAt'
            lastLogin:
              $ref: '#/components/schemas/LastLogin'
...
    UserPut:
      allOf:
      - $ref: '#/components/schemas/UserBase'
      - type: object
        properties:
          password:
            $ref: '#/components/schemas/Password'
...
    UserDisco:
      allOf:
        - $ref: '#/components/schemas/UserBase'
    UserDiscoArray:
      type: array
      items:
        $ref: '#/components/schemas/UserDisco'
    APITokenStatusEnum:
      type: string
      enum:
      - active
      - blocked
      default: "active"
      description: API token status (a `blocked` API token cannot be used).
      example: "active"
    APITokenBase:
      type: object
      properties:
        name:
          type: string
          description: API token name.
          example: "My API token name"
        status:
          $ref: '#/components/schemas/APITokenStatusEnum'
    APITokenGet:
      allOf:
        - $ref: '#/components/schemas/APITokenBase'
        - type: object
          properties:
            id:
              readOnly: true
              type: string
              format: uuid
              description: API token identifier (allocated by the platform).
              example: "a35c9fee-3893-4eb7-adde-205e1be03209"
            value:
              readOnly: true
              type: string
              description: Token to use for the signature endpoint.
              example: "b9PXEeMpSCyeaPpQiZ3Bvw=="
            createdAt:
              $ref: '#/components/schemas/CreatedAt'
            updatedAt:
              $ref: '#/components/schemas/UpdatedAt'
            deletedAt:
              $ref: '#/components/schemas/DeletedAt'
            lastUsed:
              $ref: '#/components/schemas/LastUsed'
    APITokenPost:
      allOf:
        - $ref: '#/components/schemas/APITokenBase'
        - type: object
          required:
            - name
    APITokenPut:
      allOf:
        - $ref: '#/components/schemas/APITokenBase'
    APITokenArray:
      type: array
      items:
        $ref: '#/components/schemas/APITokenGet'
    ServerConfig:
      type: object
      properties:
        identityURL:
          type: string
          example: "https://identity.mydomain.com/identity"
          description: >
            The identity URL that the server returns with a signature.
            <br>Note that the server always expects to be called on the "/identity" endpoint,
            if you want to map it to another one (like in the example, where it is linked to a dedicated subdomain)
            you must not forget it in your reverse proxy configuration.
        defaultKeyId:
          type: string
          format: uuid
          example: "7b8e5c49-18c6-4ef7-9dab-261a5e717b31"
          description: Identifier of the default key to use when signing with no user and no key specified.
        fallbackOnDefaultKey:
          type: boolean
          example: true
          description: True is the server must fallback on the default key (if any).
...
    KeyTypeEnum:
      type: string
      enum:
      - bip39
      description: Key type.
      example: "bip39"
    KeyBase:
      type: object
      properties:
        name:
          type: string
          description: Key name.
          example: "Jim Smith's key"
        status:
          $ref: '#/components/schemas/KeyStatusEnum'
    KeyGet:
      allOf:
        - $ref: '#/components/schemas/KeyBase'
        - type: object
          required:
            - name
          properties:
            id:
              readOnly: true
              type: string
              format: uuid
              description: Key identifier (allocated by the platform).
              example: "a35c9fee-3893-4eb7-adde-205e1be03209"
            pubKey:
              $ref: '#/components/schemas/PubKey'
            type:
              $ref: '#/components/schemas/KeyTypeEnum'
            expiration:
              type: integer
              format: int64
              example: 1569542400000
              description: >
                Key expiration date (Unix ms timestamp).
                <br>Note that the field is not returned if the key has no expiration date.
            expired:
              type: boolean
              example: false
              description: >
                Indicates whether the key has expired or not.
                <br>Note that the field is not returned if the key has no expiration date.
            createdAt:
              $ref: '#/components/schemas/CreatedAt'
            updatedAt:
              $ref: '#/components/schemas/UpdatedAt'
            deletedAt:
              $ref: '#/components/schemas/DeletedAt'
            lastUsed:
              $ref: '#/components/schemas/LastUsed'
...
    KeyDisco:
      allOf:
        - $ref: '#/components/schemas/KeyGet'
    KeyDiscoArray:
      type: array
      items:
        $ref: '#/components/schemas/KeyDisco'
...
    APIError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: integer
              format: int32
              example: 418
            message:
              type: string
              example: "I'm a teapot"
          required:
            - code
            - message
  parameters:
    UserIdParameter:
      name: userId
      in: path
      description: Identifier of the user.
      required: true
      schema:
        type: string
        format: uuid
        example: 'feb37e23-d04e-4e71-bf53-1f1a75ba3a68'
...
    APITokenIdParameter:
      name: APITokenId
      in: path
      description: Identifier of the API token.
      required: true
      schema:
        type: string
        format: uuid
        example: "f34d92e3-4f71-49ab-862f-69443bd48266"
...
    PubKeyParameter:
      name: pubKey
      in: path
      description: Public key (bitcoin address when using BIP39 keys).
      required: true
      schema:
        type: string
        example: "1GChJMuyxvq28F3Uksqf5v7QkxQ4WLQdBh"
    UserSearchQueryParameter:
      name: search
      in: query
      description: A string used to search users through the fields `email`, `username`, `x500CommonName`, `x500Organization` and `x500OrganizationalUnit`.
      required: true
      schema:
        type: string
        example:
          email:
            value: "john.doe@acme.com"
          username:
            value: "johndoe"
          x500CommonName:
            value: "John Doe"
          x500Organization:
            value: "Acme corp"
          x500OrganizationalUnit:
            value: "Business unit"
...
  requestBodies:
...
    ServerConfig:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ServerConfig'
      description: Replacement server config object.
      required: true
  responses:
    UserGet:
      $ref: '#/components/schemas/UserGet'
...
    UserDisco:
      $ref: '#/components/schemas/UserDisco'
    KeyDiscoArray:
      $ref: '#/components/schemas/KeyDiscoArray'
    UserDiscoArray:
      $ref: '#/components/schemas/UserDiscoArray'
    APIError:
      $ref: '#/components/schemas/APIError'
...
    APITokenGet:
      $ref: '#/components/schemas/APITokenGet'
    APITokenArray:
      $ref: '#/components/schemas/APITokenArray'
    ServerConfig:
      $ref: '#/components/schemas/ServerConfig'
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
      description: Use email/username and password to authenticate.
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Use a session cookie to authentify (see /login).
    APITokenAuth:
      type: http
      scheme: bearer
      description: Use an API token to authentify (see /api-token).
    OAuthTokenAuth:
      type: http
      scheme: bearer
      description: Use an OAuth access token to authentify.